fix i/o and hash

+++ symbol-table/src/sdbm_hash.h
@@ -2,10 +2,10 @@
 
 #include <string_view>
 
-inline unsigned int sdbm_hash(std::string_view str) {
-  unsigned int hash = 0;
-  unsigned int i = 0;
-  unsigned int len = str.length();
+inline size_t sdbm_hash(std::string_view str) {
+  size_t hash = 0;
+  size_t i = 0;
+  size_t len = str.length();
 
   for (i = 0; i < len; i++) {
     hash = (str[i]) + (hash << 6) + (hash << 16) - hash;
